Had to tweak the project and DLL bundling to make it work again.
Apparently the previous version of 3D box only worked with win10 Insider Preview build. For the regular windows necessary DLL should be bundled with the program (dxil.dll and dxcompiler.dll). There are different versions of these DLLs, use the appropriate one (for me it was "Windows Kits\10\bin\10.0.18362.0\x64")!